+++ back_end/.env
@@ -1,0 +1,4 @@
+REACT_APP_HOST=127.0.0.1
+REACT_APP_USER=root
+REACT_APP_PASSWORD=1234
+REACT_APP_DATABASE=bookDB
