+++ front_end/.env
@@ -1,0 +1,1 @@
+REACT_APP_CLIENTID="594681299704-d5o8si4lgts8kojoic1f7gktqh4a4sjl.apps.googleusercontent.com"
