+++ front_end/src/pages/PurchaseReview.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import PurchaseBanner from '../components/PurchaseBanner';
-import ReviewCreate from '../components/ReviewCreate';
+import ReviewCreate from './ReviewCreate';
 import Header from '../components/Header';
 import Footer from '../components/Footer';
